requirements-dev.txt
jsonschema>=4.20
xmlschema>=2.2
pydot>=1.4
build>=0.10
lark>=1.1
coverage>=6.3
PyYAML>=5.3
json-schema-tool==0.1.0
hypothesis-jsonschema==0.23.0
requests>=2
